child-nutrition-dashboard/requirements.txt
# Child Nutrition Dashboard Dependencies
# Streamlit framework
streamlit==1.30.0

# Database connectivity
snowflake-connector-python==3.5.0

# Data visualization
plotly==5.18.0

# AI integration
openai==1.0.0

# Data processing
pandas==2.1.0

# Environment management
python-dotenv==1.0.0

# Additional dependencies for enhanced functionality
numpy==1.24.0
requests==2.32.5
